+++ app/src/main/res/layout/activity_event.xml
@@ -14,7 +14,7 @@
         android:layout_marginStart="24dp"
         android:layout_marginTop="50dp"
         android:layout_marginBottom="72dp"
-        android:text="Event Date"
+        android:text="@string/event_date"
         app:layout_constraintBottom_toBottomOf="@+id/eventLocation"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
@@ -26,7 +26,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
         android:layout_marginTop="48dp"
-        android:text="Event Name"
+        android:text="@string/default_event_name"
         android:textSize="30sp"
         android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
@@ -40,7 +40,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
         android:layout_marginTop="294dp"
-        android:text="Event Owner"
+        android:text="@string/event_owner"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
@@ -51,7 +51,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
         android:layout_marginBottom="45dp"
-        android:text="Event Description"
+        android:text="@string/event_description"
         app:layout_constraintBottom_toTopOf="@+id/eventOwner"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
@@ -62,7 +62,7 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="24dp"
         android:layout_marginBottom="38dp"
-        android:text="Event Location"
+        android:text="@string/event_location"
         app:layout_constraintBottom_toTopOf="@+id/eventDescription"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
