+++ app/src/main/res/layout/activity_login.xml
@@ -18,13 +18,13 @@
             android:id="@+id/loginButton"
             android:layout_width="match_parent"
             android:layout_height="40dp"
-            android:text="Login" />
+            android:text="@string/login" />
 
         <Button
             android:id="@+id/registerButton"
             android:layout_width="match_parent"
             android:layout_height="40dp"
-            android:text="Register" />
+            android:text="@string/register" />
 
     </LinearLayout>
