+++ app/src/main/res/layout/activity_main.xml
@@ -12,7 +12,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="112dp"
         android:onClick="gotToProfile"
-        android:text="Profile"
+        android:text="@string/profile"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent"
@@ -26,7 +26,7 @@
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
         android:clickable="false"
-        android:text="Disconnected"
+        android:text="@string/disconnected"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -36,7 +36,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="284dp"
         android:onClick="goToMap"
-        android:text="Map"
+        android:text="@string/map"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent"
@@ -48,7 +48,7 @@
         android:layout_width="150dp"
         android:layout_height="40dp"
         android:layout_marginTop="12dp"
-        android:text="FeedView"
+        android:text="@string/feedview"
         android:textColor="@color/white"
         android:textSize="14sp"
         app:layout_constraintEnd_toEndOf="parent"
