+++ app/src/main/res/layout/content_map.xml
@@ -37,7 +37,7 @@
         android:layout_height="wrap_content"
         android:layout_alignParentBottom="true"
         android:layout_margin="16dp"
-        android:text="Switch between Events and Venues"
+        android:text="@string/switch_between_events_and_venues"
         android:textColor="#FFFFFF"
         android:background="@color/cardview_dark_background"
         android:elevation="5dp"
@@ -52,7 +52,7 @@
         android:layout_alignParentBottom="true"
         android:layout_marginStart="13dp"
         android:layout_marginBottom="90dp"
-        android:text="Zoom in"/>
+        android:text="@string/zoom_in"/>
 
     <Button
         android:id="@+id/map_settings_button"
@@ -64,7 +64,7 @@
         android:layout_marginEnd="150dp"
         android:layout_marginBottom="90dp"
         android:onClick="start_settings"
-        android:text="Settings" />
+        android:text="@string/settings" />
 
     <Button
         android:id="@+id/map_zoomout_button"
@@ -77,7 +77,7 @@
         android:layout_marginTop="13dp"
         android:layout_marginEnd="13dp"
         android:layout_marginBottom="90dp"
-        android:text="Zoom Out" />
+        android:text="@string/zoom_out" />
 
 
 </RelativeLayout>
