+++ app/src/main/res/layout/event_page.xml
@@ -2,8 +2,6 @@
 
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
 
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -31,7 +29,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="8dp"
-        android:text="Name"
+        android:text="@string/name"
         android:textColor="#000"
         android:textSize="22sp"
         android:textStyle="bold" />
@@ -41,7 +39,7 @@
         android:id="@+id/tag"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Tag"
+        android:text="@string/tag"
         android:textColor="#A5A5A5"
         android:textSize="18sp"
         android:textStyle="bold" />
