+++ app/src/main/res/layout/event_row.xml
@@ -16,7 +16,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="Event"
+            android:text="@string/event"
             android:id="@+id/event"
             android:layout_alignParentTop="true"
             android:textAlignment="center" />
@@ -25,7 +25,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:textAppearance="?android:attr/textAppearanceMedium"
-            android:text="ID"
+            android:text="@string/id"
             android:id="@+id/id"
             android:layout_below="@+id/event"
             android:textAlignment="center" />
@@ -33,7 +33,7 @@
         <TextView
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:text="Data"
+            android:text="@string/data"
             android:id="@+id/data"
             android:textAlignment="center"
             android:layout_marginTop="8dp" />
