+++ app/src/main/res/values/strings.xml
@@ -17,7 +17,6 @@
     values-round/strings.xml for round devices.
     -->
     <string name="hello_world">Hello Square World!</string>
-    <string name="event_name">Event Name</string>
     <string name="default_event_name">Event Name</string>
     <string name="default_event_description">Event Description</string>
     <string name="default_validate_button">Validate Selection</string>
@@ -42,4 +41,35 @@
     <string name="settings_distance_slider">Max distance = </string>
     <string name="settings_accept_button">Accept</string>
     <string name="settings_discard_button">Discard</string>
+    <string name="event_date">Event Date</string>
+    <string name="event_owner">Event Owner</string>
+    <string name="event_description">Event Description</string>
+    <string name="event_location">Event Location</string>
+    <string name="profile">Profile</string>
+    <string name="disconnected">Disconnected</string>
+    <string name="map">Map</string>
+    <string name="feedview">FeedView</string>
+    <string name="switch_between_events_and_venues">Switch between Events and Venues</string>
+    <string name="zoom_in">Zoom in</string>
+    <string name="settings">Settings</string>
+    <string name="zoom_out">Zoom Out</string>
+    <string name="welcome_to_your_profile">Welcome to your Profile</string>
+    <string name="log_out">Log Out</string>
+    <string name="email_address">Email Address</string>
+    <string name="firstname">Firstname...</string>
+    <string name="lastname">Lastname...</string>
+    <string name="date_of_birth">Date of birth...</string>
+    <string name="save">Save</string>
+    <string name="name">Name</string>
+    <string name="tag">Tag</string>
+    <string name="event">Event</string>
+    <string name="id">ID</string>
+    <string name="data">Data</string>
+    <string name="login">Login</string>
+    <string name="email_hint">Email...</string>
+    <string name="confirm_password">Confirm Password</string>
+    <string name="register">Register</string>
+    <string name="registration">Registration</string>
+    <string name="password_hint">Password...</string>
+    <string name="i_m_an_organizer">I\'m an organizer</string>
 </resources>
